+++ app/src/features/approval/index.ts
@@ -1,3 +1,4 @@
 export { ApprovalQueuePage } from './pages/ApprovalQueuePage'
 export { approvalStore, useApprovalStore } from './store/approvalStore'
 export * from './types/approvalTypes'
+export * from './utils/approvalInsights'